@@ -8,10 +8,9 @@
   },
   "main": "index.js",
   "scripts": {
-    "prepublish": "rm -rf lib/* && node_modules/.bin/gulp build",
-    "test": "node_modules/.bin/gulp test",
-    "test-ci": "node_modules/.bin/gulp test-ci",
-    "coverage": "./node_modules/coveralls/bin/coveralls.js < coverage/lcov.info"
+    "test": "nyc --reporter=html --reporter=text gulp test",
+    "test:ci": "gulp test-ci",
+    "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "author": "SoapBox Innovations (@SoapBoxHQ)",
   "license": "MIT",
@@ -28,27 +27,28 @@
     "mentions"
   ],
   "devDependencies": {
-    "amd-optimize": "^0.6.1",
+    "babel-core": "^6.26.0",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-es2015-modules-amd": "^6.24.1",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "chai": "^3.5.0",
+    "chai": "^4.1.2",
     "coveralls": "^3.0.0",
     "glob": "^7.1.2",
     "gulp": "^3.8.11",
-    "gulp-babel": "^6.1.2",
+    "gulp-babel": "^7.0.1",
     "gulp-clean": "^0.3.2",
+    "gulp-cli": "^2.0.1",
     "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^3.0.1",
-    "gulp-istanbul": "^1.1.2",
-    "gulp-mocha": "^3.0.1",
+    "gulp-eslint": "^4.0.2",
+    "gulp-mocha": "^5.0.0",
     "gulp-rename": "^1.2.0",
     "gulp-replace": "^0.5.4",
     "gulp-uglify": "^1.5.4",
     "gulp-util": "^3.0.7",
     "gulp-wrap": "^0.13.0",
-    "jsdom": "^9.12.0",
-    "karma": "^1.7.1",
+    "jsdom": "^11.6.1",
+    "karma": "^2.0.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-firefox-launcher": "^1.0.1",
     "karma-phantomjs-launcher": "^1.0.4",
@@ -57,18 +57,19 @@
     "lazypipe": "^1.0.1",
     "lodash": "^4.11.1",
     "merge-stream": "^1.0.0",
-    "mocha": "^3.5.3",
+    "mocha": "^5.0.0",
+    "nyc": "^11.4.1",
     "phantomjs-prebuilt": "^2.1.15",
     "qunitjs": "^1.14.0",
     "requirejs": "^2.3.5",
     "rollup": "~0.40.0",
-    "run-sequence": "^1.1.5",
+    "run-sequence": "^2.2.1",
     "simple-html-tokenizer": "git+https://github.com/nfrasser/simple-html-tokenizer.git#04799f4638ec5ed903a4e5aa6e832269fa59be6b",
     "through2": "^2.0.1"
   },
   "optionalDependencies": {
-    "jquery": ">=1.9.0",
-    "react": ">=0.14.0",
-    "react-dom": ">=0.14.0"
+    "jquery": "^3.3.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
   }
 }
