@@ -1,8 +1,8 @@
 import HTML5Tokenizer from './simple-html-tokenizer';
 import * as linkify from './linkify';
 
-const {options} = linkify;
-const {Options} = options;
+const { options } = linkify;
+const { Options } = options;
 
 const StartTag = 'StartTag';
 const EndTag = 'EndTag';
@@ -13,7 +13,7 @@ const Comment = 'Comment';
 	`tokens` and `token` in this section refer to tokens generated by the HTML
 	parser.
 */
-export default function linkifyHtml(str, opts={}) {
+export default function linkifyHtml(str, opts = {}) {
 	let tokens = HTML5Tokenizer.tokenize(str);
 	let linkifiedTokens = [];
 	let linkified = [];
@@ -54,7 +54,7 @@ export default function linkifyHtml(str, opts={}) {
 	for (i = 0; i < linkifiedTokens.length; i++) {
 		let token = linkifiedTokens[i];
 		switch (token.type) {
-		case StartTag:
+		case StartTag: {
 			let link = '<' + token.tagName;
 			if (token.attributes.length > 0) {
 				let attrs = attrsToStrings(token.attributes);
@@ -63,6 +63,7 @@ export default function linkifyHtml(str, opts={}) {
 			link += '>';
 			linkified.push(link);
 			break;
+		}
 		case EndTag:
 			linkified.push(`</${token.tagName}>`);
 			break;
@@ -103,7 +104,6 @@ function linkifyChars(str, opts) {
 		}
 
 		let {
-			href,
 			formatted,
 			formattedHref,
 			tagName,
