@@ -1,7 +1,7 @@
 // React expects to be loaded asynchronously by AMD
-define('react', ['base/node_modules/react/dist/react', 'exports'], function (React, exports) {
+define('react', ['base/vendor/react.min', 'exports'], function (React, exports) {
 	return exports.React = React;
 });
-define('react-dom', ['base/node_modules/react-dom/dist/react-dom', 'exports'], function (ReactDOM, exports) {
+define('react-dom', ['base/vendor/react-dom.min', 'exports'], function (ReactDOM, exports) {
 	return exports.ReactDOM = ReactDOM;
 });
